Day_2/requirements.txt
google-adk
python-docx>=1.1.2
requests>=2.31.0
beautifulsoup4>=4.12.0
requests
pandas
openpyxl
python-dotenv